backend/requirements.txt
## System requirements
pip>=9.0
setuptools>=8.0

## Django 1.10+ and utils
Django~=2.0.5
docutils>=0.12

## Django-cors-headers to allow cross site references
django-cors-headers==1.3.1

## django-gravatar for Gravatar support
django-gravatar2==1.4.0
djangorestframework>=3.8
reportlab>=3.3.0

#Postgresql driver
psycopg2>=2.7.3.2
